Add option to write env vars to spark-env
This commit adds support for adding an arbitrary set
of environment variables to `spark-env.sh` by passing
in a list of objects with `key` and `value` attributes
by overriding the `spark_env_extras` variable when a
template inherits this template.

For example, if spark is running on top of a mesos cluster,
users of the role can override `spark_env_extras` to set the
path to the mesos native library.

```
spark_env_extras:
  MESOS_NATIVE_LIBRARY: "/usr/local/lib/libmesos.so"
```